refactor: split core metrics into dedicated modules
