+++ projects/04-llm-adapter/adapter/core/parallel_state.py
@@ -10,7 +10,7 @@ from uuid import uuid4
 
 from .config import ProviderConfig
 from .datasets import GoldenTask
-from .metrics import BudgetSnapshot, EvalMetrics, now_ts, RunMetrics
+from .metrics.models import BudgetSnapshot, EvalMetrics, RunMetrics, now_ts
 from .providers import BaseProvider
 
 if TYPE_CHECKING:  # pragma: no cover - 型補完用
